合入apache master

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/PrivateCellUtil.java
@@ -132,6 +132,7 @@ public final class PrivateCellUtil {
     private static final int HEAP_SIZE_OVERHEAD = ClassSize.OBJECT + 2 * ClassSize.REFERENCE;
 
     /**
+     * Construct a TagRewriteCell
      * @param cell The original Cell which it rewrites
      * @param tags the tags bytes. The array suppose to contain the tags bytes alone.
      */
@@ -828,8 +829,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * @return True if a delete type, a {@link KeyValue.Type#Delete} or a {KeyValue.Type#DeleteFamily}
-   *         or a {@link KeyValue.Type#DeleteColumn} KeyValue type.
+   * Return true if a delete type, a {@link KeyValue.Type#Delete} or a {KeyValue.Type#DeleteFamily}
+   * or a {@link KeyValue.Type#DeleteColumn} KeyValue type.
    */
   public static boolean isDelete(final byte type) {
     return KeyValue.Type.Delete.getCode() <= type && type <= KeyValue.Type.DeleteFamily.getCode();
@@ -868,9 +869,7 @@ public final class PrivateCellUtil {
     return output;
   }
 
-  /**
-   * Copies the tags info into the tag portion of the cell nnn * @return position after tags
-   */
+  /** Copies the tags info into the tag portion of the cell */
   public static int copyTagsTo(Cell cell, byte[] destination, int destinationOffset) {
     int tlen = cell.getTagsLength();
     if (cell instanceof ByteBufferExtendedCell) {
@@ -884,9 +883,7 @@ public final class PrivateCellUtil {
     return destinationOffset + tlen;
   }
 
-  /**
-   * Copies the tags info into the tag portion of the cell nnn * @return the position after tags
-   */
+  /** Copies the tags info into the tag portion of the cell */
   public static int copyTagsTo(Cell cell, ByteBuffer destination, int destinationOffset) {
     int tlen = cell.getTagsLength();
     if (cell instanceof ByteBufferExtendedCell) {
@@ -900,6 +897,7 @@ public final class PrivateCellUtil {
   }
 
   /**
+   * Return tags in the given Cell as a List
    * @param cell The Cell
    * @return Tags in the given Cell as a List
    */
@@ -945,7 +943,7 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Util method to iterate through the tags in the given cell.
+   * Utility method to iterate through the tags in the given cell.
    * @param cell The Cell over which tags iterator is needed.
    * @return iterator for the tags
    */
@@ -1028,9 +1026,7 @@ public final class PrivateCellUtil {
       && (end1.length == 0 || start2.length == 0 || Bytes.compareTo(start2, end1) < 0);
   }
 
-  /**
-   * Write rowkey excluding the common part. nnnnn
-   */
+  /** Write rowkey excluding the common part. */
   public static void writeRowKeyExcludingCommon(Cell cell, short rLen, int commonPrefix,
     DataOutputStream out) throws IOException {
     if (commonPrefix == 0) {
@@ -1050,7 +1046,7 @@ public final class PrivateCellUtil {
    * Writes the row from the given cell to the output stream excluding the common prefix
    * @param out     The dataoutputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param rlength the row length n
+   * @param rlength the row length
    */
   public static void writeRowSkippingBytes(DataOutputStream out, Cell cell, short rlength,
     int commonPrefix) throws IOException {
@@ -1238,8 +1234,8 @@ public final class PrivateCellUtil {
 
   /**
    * Compares only the key portion of a cell. It does not include the sequence id/mvcc of the cell
-   * nn * @return an int greater than 0 if left &gt; than right lesser than 0 if left &lt; than
-   * right equal to 0 if left is equal to right
+   * @return an int greater than 0 if left &gt; than right lesser than 0 if left &lt; than right
+   *         equal to 0 if left is equal to right
    */
   public static final int compareKeyIgnoresMvcc(CellComparator comparator, Cell left, Cell right) {
     return ((CellComparatorImpl) comparator).compare(left, right, true);
@@ -1834,7 +1830,7 @@ public final class PrivateCellUtil {
 
   private static class FirstOnRowColCell extends FirstOnRowCell {
     // @formatter:off
-    private static final long FIXED_HEAPSIZE = FirstOnRowCell.FIXED_HEAPSIZE
+    private static final long FIXED_HEAPSIZE = (long) FirstOnRowCell.FIXED_HEAPSIZE
       + Bytes.SIZEOF_BYTE // flength
       + Bytes.SIZEOF_INT * 3 // foffset, qoffset, qlength
       + ClassSize.REFERENCE * 2; // fArray, qArray
@@ -2003,7 +1999,7 @@ public final class PrivateCellUtil {
 
   private static class LastOnRowColCell extends LastOnRowCell {
     // @formatter:off
-    private static final long FIXED_OVERHEAD = LastOnRowCell.FIXED_OVERHEAD
+    private static final long FIXED_OVERHEAD = (long) LastOnRowCell.FIXED_OVERHEAD
       + ClassSize.REFERENCE * 2 // fArray and qArray
       + Bytes.SIZEOF_INT * 3 // foffset, qoffset, qlength
       + Bytes.SIZEOF_BYTE; // flength
@@ -2198,7 +2194,7 @@ public final class PrivateCellUtil {
   /**
    * Writes the Cell's key part as it would have serialized in a KeyValue. The format is &lt;2 bytes
    * rk len&gt;&lt;rk&gt;&lt;1 byte cf len&gt;&lt;cf&gt;&lt;qualifier&gt;&lt;8 bytes
-   * timestamp&gt;&lt;1 byte type&gt; nnn
+   * timestamp&gt;&lt;1 byte type&gt;
    */
   public static void writeFlatKey(Cell cell, DataOutput out) throws IOException {
     short rowLen = cell.getRowLength();
@@ -2230,7 +2226,7 @@ public final class PrivateCellUtil {
   /**
    * Deep clones the given cell if the cell supports deep cloning
    * @param cell the cell to be cloned
-   * @return the cloned cell n
+   * @return the cloned cell
    */
   public static Cell deepClone(Cell cell) throws CloneNotSupportedException {
     if (cell instanceof ExtendedCell) {
@@ -2244,7 +2240,7 @@ public final class PrivateCellUtil {
    * @param cell     the cell to be written
    * @param out      the outputstream
    * @param withTags if tags are to be written or not
-   * @return the total bytes written n
+   * @return the total bytes written
    */
   public static int writeCell(Cell cell, OutputStream out, boolean withTags) throws IOException {
     if (cell instanceof ExtendedCell) {
@@ -2319,8 +2315,8 @@ public final class PrivateCellUtil {
 
   /**
    * Sets the given seqId to the cell. Marked as audience Private as of 1.2.0. Setting a Cell
-   * sequenceid is an internal implementation detail not for general public use. nn * @throws
-   * IOException when the passed cell is not of type {@link ExtendedCell}
+   * sequenceid is an internal implementation detail not for general public use.
+   * @throws IOException when the passed cell is not of type {@link ExtendedCell}
    */
   public static void setSequenceId(Cell cell, long seqId) throws IOException {
     if (cell instanceof ExtendedCell) {
@@ -2332,8 +2328,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Sets the given timestamp to the cell. nn * @throws IOException when the passed cell is not of
-   * type {@link ExtendedCell}
+   * Sets the given timestamp to the cell.
+   * @throws IOException when the passed cell is not of type {@link ExtendedCell}
    */
   public static void setTimestamp(Cell cell, long ts) throws IOException {
     if (cell instanceof ExtendedCell) {
@@ -2345,7 +2341,7 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Sets the given timestamp to the cell. n * @param ts buffer containing the timestamp value
+   * Sets the given timestamp to the cell.
    * @throws IOException when the passed cell is not of type {@link ExtendedCell}
    */
   public static void setTimestamp(Cell cell, byte[] ts) throws IOException {
@@ -2359,7 +2355,8 @@ public final class PrivateCellUtil {
 
   /**
    * Sets the given timestamp to the cell iff current timestamp is
-   * {@link HConstants#LATEST_TIMESTAMP}. nn * @return True if cell timestamp is modified.
+   * {@link HConstants#LATEST_TIMESTAMP}.
+   * @return True if cell timestamp is modified.
    * @throws IOException when the passed cell is not of type {@link ExtendedCell}
    */
   public static boolean updateLatestStamp(Cell cell, long ts) throws IOException {
@@ -2372,7 +2369,7 @@ public final class PrivateCellUtil {
 
   /**
    * Sets the given timestamp to the cell iff current timestamp is
-   * {@link HConstants#LATEST_TIMESTAMP}. n * @param ts buffer containing the timestamp value
+   * {@link HConstants#LATEST_TIMESTAMP}.
    * @return True if cell timestamp is modified.
    * @throws IOException when the passed cell is not of type {@link ExtendedCell}
    */
@@ -2388,7 +2385,7 @@ public final class PrivateCellUtil {
    * Writes the row from the given cell to the output stream
    * @param out     The outputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param rlength the row length n
+   * @param rlength the row length
    */
   public static void writeRow(OutputStream out, Cell cell, short rlength) throws IOException {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2403,7 +2400,7 @@ public final class PrivateCellUtil {
    * Writes the family from the given cell to the output stream
    * @param out     The outputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param flength the family length n
+   * @param flength the family length
    */
   public static void writeFamily(OutputStream out, Cell cell, byte flength) throws IOException {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2418,7 +2415,7 @@ public final class PrivateCellUtil {
    * Writes the qualifier from the given cell to the output stream
    * @param out     The outputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param qlength the qualifier length n
+   * @param qlength the qualifier length
    */
   public static void writeQualifier(OutputStream out, Cell cell, int qlength) throws IOException {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2434,7 +2431,7 @@ public final class PrivateCellUtil {
    * Writes the qualifier from the given cell to the output stream excluding the common prefix
    * @param out     The dataoutputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param qlength the qualifier length n
+   * @param qlength the qualifier length
    */
   public static void writeQualifierSkippingBytes(DataOutputStream out, Cell cell, int qlength,
     int commonPrefix) throws IOException {
@@ -2453,7 +2450,7 @@ public final class PrivateCellUtil {
    * Writes the value from the given cell to the output stream
    * @param out     The outputstream to which the data has to be written
    * @param cell    The cell whose contents has to be written
-   * @param vlength the value length n
+   * @param vlength the value length
    */
   public static void writeValue(OutputStream out, Cell cell, int vlength) throws IOException {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2468,7 +2465,7 @@ public final class PrivateCellUtil {
    * Writes the tag from the given cell to the output stream
    * @param out        The outputstream to which the data has to be written
    * @param cell       The cell whose contents has to be written
-   * @param tagsLength the tag length n
+   * @param tagsLength the tag length
    */
   public static void writeTags(OutputStream out, Cell cell, int tagsLength) throws IOException {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2501,7 +2498,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Converts the rowkey bytes of the given cell into an int value n * @return rowkey as int
+   * Converts the rowkey bytes of the given cell into an int value
+   * @return rowkey as int
    */
   public static int getRowAsInt(Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2512,7 +2510,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Converts the value bytes of the given cell into a long value n * @return value as long
+   * Converts the value bytes of the given cell into a long value
+   * @return value as long
    */
   public static long getValueAsLong(Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2523,7 +2522,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Converts the value bytes of the given cell into a int value n * @return value as int
+   * Converts the value bytes of the given cell into a int value
+   * @return value as int
    */
   public static int getValueAsInt(Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2534,7 +2534,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Converts the value bytes of the given cell into a double value n * @return value as double
+   * Converts the value bytes of the given cell into a double value
+   * @return value as double
    */
   public static double getValueAsDouble(Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2545,7 +2546,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Converts the value bytes of the given cell into a BigDecimal n * @return value as BigDecimal
+   * Converts the value bytes of the given cell into a BigDecimal
+   * @return value as BigDecimal
    */
   public static BigDecimal getValueAsBigDecimal(Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2711,8 +2713,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * @return An new cell is located following input cell. If both of type and timestamp are minimum,
-   *         the input cell will be returned directly.
+   * Return a new cell is located following input cell. If both of type and timestamp are minimum,
+   * the input cell will be returned directly.
    */
   public static Cell createNextOnRowCol(Cell cell) {
     long ts = cell.getTimestamp();
@@ -2766,8 +2768,9 @@ public final class PrivateCellUtil {
   /**
    * Estimate based on keyvalue's serialization format in the RPC layer. Note that there is an extra
    * SIZEOF_INT added to the size here that indicates the actual length of the cell for cases where
-   * cell's are serialized in a contiguous format (For eg in RPCs). n * @return Estimate of the
-   * <code>cell</code> size in bytes plus an extra SIZEOF_INT indicating the actual cell length.
+   * cell's are serialized in a contiguous format (For eg in RPCs).
+   * @return Estimate of the <code>cell</code> size in bytes plus an extra SIZEOF_INT indicating the
+   *         actual cell length.
    */
   public static int estimatedSerializedSizeOf(final Cell cell) {
     return cell.getSerializedSize() + Bytes.SIZEOF_INT;
@@ -2787,9 +2790,9 @@ public final class PrivateCellUtil {
   /**
    * This method exists just to encapsulate how we serialize keys. To be replaced by a factory that
    * we query to figure what the Cell implementation is and then, what serialization engine to use
-   * and further, how to serialize the key for inclusion in hfile index. TODO. n * @return The key
-   * portion of the Cell serialized in the old-school KeyValue way or null if passed a null
-   * <code>cell</code>
+   * and further, how to serialize the key for inclusion in hfile index. TODO.
+   * @return The key portion of the Cell serialized in the old-school KeyValue way or null if passed
+   *         a null <code>cell</code>
    */
   public static byte[] getCellKeySerializedAsKeyValueKey(final Cell cell) {
     if (cell == null) return null;
@@ -2799,8 +2802,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Create a Cell that is smaller than all other possible Cells for the given Cell's row. n
-   * * @return First possible Cell on passed Cell's row.
+   * Create a Cell that is smaller than all other possible Cells for the given Cell's row.
+   * @return First possible Cell on passed Cell's row.
    */
   public static Cell createFirstOnRow(final Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2864,8 +2867,8 @@ public final class PrivateCellUtil {
 
   /**
    * Create a Cell that is smaller than all other possible Cells for the given Cell's rk:cf and
-   * passed qualifier. nnnn * @return Last possible Cell on passed Cell's rk:cf and passed
-   * qualifier.
+   * passed qualifier.
+   * @return Last possible Cell on passed Cell's rk:cf and passed qualifier.
    */
   public static Cell createFirstOnRowCol(final Cell cell, byte[] qArray, int qoffest, int qlength) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2885,7 +2888,7 @@ public final class PrivateCellUtil {
    * Creates the first cell with the row/family/qualifier of this cell and the given timestamp. Uses
    * the "maximum" type that guarantees that the new cell is the lowest possible for this
    * combination of row, family, qualifier, and timestamp. This cell's own timestamp is ignored.
-   * @param cell - cell n
+   * @param cell - cell
    */
   public static Cell createFirstOnRowColTS(Cell cell, long ts) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2903,8 +2906,8 @@ public final class PrivateCellUtil {
   }
 
   /**
-   * Create a Cell that is larger than all other possible Cells for the given Cell's row. n
-   * * @return Last possible Cell on passed Cell's row.
+   * Create a Cell that is larger than all other possible Cells for the given Cell's row.
+   * @return Last possible Cell on passed Cell's row.
    */
   public static Cell createLastOnRow(final Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
@@ -2921,7 +2924,8 @@ public final class PrivateCellUtil {
   /**
    * Create a Cell that is larger than all other possible Cells for the given Cell's rk:cf:q. Used
    * in creating "fake keys" for the multi-column Bloom filter optimization to skip the row/column
-   * we already know is not in the file. n * @return Last possible Cell on passed Cell's rk:cf:q.
+   * we already know is not in the file.
+   * @return Last possible Cell on passed Cell's rk:cf:q.
    */
   public static Cell createLastOnRowCol(final Cell cell) {
     if (cell instanceof ByteBufferExtendedCell) {
